HID: primax: fix wireless keyboard incorrect report for keystrokes.
Primax wireless keyboard/mouse kit shipped with Acer Desktop/All-In-One
reports each keystroke with LED hid usage which is supposed to be
keyboard hid usage. This commit correct the wrong HID_USAGE_PAGE and lets
the standard code to do the rest.

We do not have model info for the 4e80 device.
The 4e57 device engineering sample has a barcode sticker that mentions
Primax, however the keyboard itself is branded as Acer and the
model code printed underneath is KBRF96211.

Neither keyboard has any LEDs.

The same devices also come with a wireless mouse, we are careful to
only apply this quirk to the keyboard to avoid interference with
mouse input.

Signed-off-by: Chris Chiu <chiu@endlessm.com>

https://phabricator.endlessm.com/T14394
https://phabricator.endlessm.com/T19434
https://phabricator.endlessm.com/T19771