drm/nouveau: Disable runtime pm on Asus Notebooks with NV GM108
The Asus X542UF/X542UQ, P1440UF/P2440UF and more models with NVIDIA
940MX, MX110/MX130 has a problem with the interaction between runtime
pm and suspend/resume which leads to the machine hanging on boot or
shutdown. This commit disables runtime pm for all Asus Notebooks with
the display chip NV GM108.

It has to be reverted once if there's a upstream fix.

https://phabricator.endlessm.com/T20646

Signed-off-by: Chris Chiu <chiu@endlessm.com>